+++ test/clients/run-client.test.ts
@@ -85,14 +85,27 @@ describe('run-client', () => {
             expect(updateRunSpy).toHaveBeenCalledWith('test-run', mockRun);
         });
 
-        it('declares a Run lost if not found', () => {
-            // TODO: test
+        it('declares a Run lost if not found', async () => {
+            const getSpy = vi.spyOn(RunClient.prototype, 'get')
+                .mockImplementation(async () => null);
+            const declareLostRunSpy = vi.spyOn(RunsTracker.prototype, 'declareLostRun');
+            const run = await runClient.get();
+            expect(run).toEqual(null);
+            expect(getSpy).toHaveBeenCalledTimes(1);
+            expect(declareLostRunSpy).toHaveBeenCalledTimes(1);
+            expect(declareLostRunSpy).toHaveBeenCalledWith('test-run', 'Actor client could not retrieve the Run');
         });
     });
 
     describe('abort', () => {
-        it('updates the tracker when called', () => {
-            // TODO: test
+        it('updates the tracker when called', async () => {
+            const abortSpy = vi.spyOn(RunClient.prototype, 'abort')
+                .mockImplementation(async () => mockRun);
+            const run = await runClient.abort();
+            expect(run).toEqual(mockRun);
+            expect(abortSpy).toHaveBeenCalledTimes(1);
+            expect(updateRunSpy).toHaveBeenCalledTimes(2); // start + get
+            expect(updateRunSpy).toHaveBeenCalledWith('test-run', mockRun);
         });
     });
 
@@ -109,26 +122,50 @@ describe('run-client', () => {
     });
 
     describe('reboot', () => {
-        it('updates the tracker when called', () => {
-            // TODO: test
+        it('updates the tracker when called', async () => {
+            const rebootSpy = vi.spyOn(RunClient.prototype, 'reboot')
+                .mockImplementation(async () => mockRun);
+            const run = await runClient.reboot();
+            expect(run).toEqual(mockRun);
+            expect(rebootSpy).toHaveBeenCalledTimes(1);
+            expect(updateRunSpy).toHaveBeenCalledTimes(2); // start + get
+            expect(updateRunSpy).toHaveBeenCalledWith('test-run', mockRun);
         });
     });
 
     describe('update', () => {
-        it('updates the tracker when called', () => {
-            // TODO: test
+        it('updates the tracker when called', async () => {
+            const updateSpy = vi.spyOn(RunClient.prototype, 'update')
+                .mockImplementation(async () => mockRun);
+            const run = await runClient.update({ statusMessage: 'test' });
+            expect(run).toEqual(mockRun);
+            expect(updateSpy).toHaveBeenCalledTimes(1);
+            expect(updateRunSpy).toHaveBeenCalledTimes(2); // start + get
+            expect(updateRunSpy).toHaveBeenCalledWith('test-run', mockRun);
         });
     });
 
     describe('resurrect', () => {
-        it('updates the tracker when called', () => {
-            // TODO: test
+        it('updates the tracker when called', async () => {
+            const resurrectSpy = vi.spyOn(RunClient.prototype, 'resurrect')
+                .mockImplementation(async () => mockRun);
+            const run = await runClient.resurrect();
+            expect(run).toEqual(mockRun);
+            expect(resurrectSpy).toHaveBeenCalledTimes(1);
+            expect(updateRunSpy).toHaveBeenCalledTimes(2); // start + get
+            expect(updateRunSpy).toHaveBeenCalledWith('test-run', mockRun);
         });
     });
 
     describe('waitForFinish', () => {
-        it('updates the tracker when called', () => {
-            // TODO: test
+        it('updates the tracker when called', async () => {
+            const waitForFinishSpy = vi.spyOn(RunClient.prototype, 'waitForFinish')
+                .mockImplementation(async () => mockRun);
+            const run = await runClient.waitForFinish();
+            expect(run).toEqual(mockRun);
+            expect(waitForFinishSpy).toHaveBeenCalledTimes(1);
+            expect(updateRunSpy).toHaveBeenCalledTimes(2); // start + get
+            expect(updateRunSpy).toHaveBeenCalledWith('test-run', mockRun);
         });
     });
 });